Move testInsertWithCoercion to engine-only tests
The coercion during insert is applied exclusively by the engine, so
there currently is no need to exercise the test against connectors.
