+++ presto-accumulo/src/test/java/io/prestosql/plugin/accumulo/TestAccumuloDistributedQueries.java
@@ -161,12 +161,6 @@ public class TestAccumuloDistributedQueries
         assertUpdate("DROP TABLE test_insert");
     }
 
-    @Override
-    public void testInsertWithCoercion()
-    {
-        // Override because of non-canonical varchar mapping
-    }
-
     @Override // Overridden because we currently do not support arrays with null elements
     public void testInsertArray()
     {
