+++ presto-testing/src/main/java/io/prestosql/testing/AbstractTestDistributedQueries.java
@@ -551,43 +551,6 @@ public abstract class AbstractTestDistributedQueries
     }
 
     @Test
-    public void testInsertWithCoercion()
-    {
-        String tableName = "test_insert_with_coercion_" + randomTableSuffix();
-
-        assertUpdate("CREATE TABLE " + tableName + " (" +
-                "tinyint_column TINYINT, " +
-                "integer_column INTEGER, " +
-                "decimal_column DECIMAL(5, 3), " +
-                "real_column REAL, " +
-                "char_column CHAR(3), " +
-                "bounded_varchar_column VARCHAR(3), " +
-                "unbounded_varchar_column VARCHAR, " +
-                "date_column DATE)");
-
-        assertUpdate("INSERT INTO " + tableName + " (tinyint_column, integer_column, decimal_column, real_column) VALUES (1e0, 2e0, 3e0, 4e0)", 1);
-        assertUpdate("INSERT INTO " + tableName + " (char_column, bounded_varchar_column, unbounded_varchar_column) VALUES (CAST('aa     ' AS varchar), CAST('aa     ' AS varchar), CAST('aa     ' AS varchar))", 1);
-        assertUpdate("INSERT INTO " + tableName + " (char_column, bounded_varchar_column, unbounded_varchar_column) VALUES (NULL, NULL, NULL)", 1);
-        assertUpdate("INSERT INTO " + tableName + " (char_column, bounded_varchar_column, unbounded_varchar_column) VALUES (CAST(NULL AS varchar), CAST(NULL AS varchar), CAST(NULL AS varchar))", 1);
-        assertUpdate("INSERT INTO " + tableName + " (date_column) VALUES (TIMESTAMP '2019-11-18 22:13:40')", 1);
-
-        assertQuery(
-                "SELECT * FROM " + tableName,
-                "VALUES " +
-                        "(1, 2, 3, 4, NULL, NULL, NULL, NULL), " +
-                        "(NULL, NULL, NULL, NULL, 'aa ', 'aa ', 'aa     ', NULL), " +
-                        "(NULL, NULL, NULL, NULL, NULL, NULL, NULL, NULL), " +
-                        "(NULL, NULL, NULL, NULL, NULL, NULL, NULL, NULL), " +
-                        "(NULL, NULL, NULL, NULL, NULL, NULL, NULL, DATE '2019-11-18')");
-
-        assertQueryFails("INSERT INTO " + tableName + " (integer_column) VALUES (3e9)", "Out of range for integer: 3.0E9");
-        assertQueryFails("INSERT INTO " + tableName + " (char_column) VALUES ('abcd')", "\\QCannot truncate non-space characters when casting from varchar(4) to char(3) on INSERT");
-        assertQueryFails("INSERT INTO " + tableName + " (bounded_varchar_column) VALUES ('abcd')", "\\QCannot truncate non-space characters when casting from varchar(4) to varchar(3) on INSERT");
-
-        assertUpdate("DROP TABLE " + tableName);
-    }
-
-    @Test
     public void testInsertUnicode()
     {
         String tableName = "test_insert_unicode_" + randomTableSuffix();
